refactor(settings): consolidate advanced settings form logic in edit_page
- Moved advanced settings form items from source_provider.dart to edit_page.dart
- Created AdvancedSettingsTile widget component
- Added APK extension filtering utilities
- Removed duplicate regExValidator function
- Updated app source files to import edit_page.dart
- Added linter rules for implementation_imports and build context safety
