+++ analysis_options.yaml
@@ -11,7 +11,9 @@ analyzer:
   errors:
     constant_identifier_names: ignore
     file_names: ignore
+    implementation_imports: ignore
     non_constant_identifier_names: ignore
+    use_build_context_synchronously: ignore
 include: package:flutter_lints/flutter.yaml
 
 linter:
