+++ lib/app_sources/html.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:updatium/pages/edit_page.dart';
 import 'package:updatium/services/slang_converter.dart';
 import 'package:html/parser.dart';
 import 'package:http/http.dart';
